==> R_analysis/01_Table1_summary_statistics.Rmd <==
---
title: "Table 1. Summary characteristics of the study groups."
output:
  html_document:
    df_print: paged
  word_document: default
  pdf_document: default
---

```{r, include = FALSE}

## R Markdown
library(dplyr)
library(kableExtra)
library(rmarkdown)
set.seed(42)
library(qwraps2)

# define the markup language we are working in.
# options(qwraps2_markup = "latex") is also supported.
options(qwraps2_markup = "markdown")


metadata <- read.csv(file = "metadata_files/cohort_metadata_combined.csv")

metadata %>% 
  group_by(diagnosis_last_record=="CD", Montreal_L) %>%
  summarise(n()) #Type of CD in cohort

metadata$age_range <- ifelse(metadata$age < 41, "18-40",
                             ifelse(metadata$age > 40 & metadata$age < 61, "41-60", 
                                    ifelse(metadata$age > 60 & metadata$age < 81, "61-80", 
                                           ifelse(metadata$age > 80, "81+", NA))))

table_summary <-
  list("Sex" =
         list("Female"    = ~ n_perc(sex=="Female"),
              "Male"      = ~ n_perc(sex=="Male"),
              "Unspecified" = ~ n_perc(sex=="")),
       "Age (years)" =
         list("Median (IQR)" = ~ median_iqr(age, na_rm=T, show_n = "never"),
              "18-40" = ~ n_perc(age_range=="18-40", na_rm=T, show_denom = "never"),
              "41-60" = ~ n_perc(age_range=="41-60", na_rm=T, show_denom = "never"),
              "61-80" = ~ n_perc(age_range=="61-80", na_rm=T, show_denom = "never"),
              "81+" = ~ n_perc(age_range=="81+", na_rm=T, show_denom = "never"),
              "Missing" = ~sum(is.na(age_range))),
       "BMI" =
         list("Median (IQR)" = ~ median_iqr(bmi, na_rm=T, show_n = "never"),
              "Missing" = ~sum(is.na(bmi))),
              "Smoking status" =
         list("Current" = ~n_perc(smoking_status=="current", na_rm=T, show_denom = "never"),
              "Past"  = ~n_perc(smoking_status=="past", na_rm=T, show_denom = "never"),
              "Never" = ~n_perc(smoking_status=="never", na_rm=T, show_denom = "never"),
              "Missing" = ~sum(is.na(smoking_status))),
       "Diagnosis" =
         list("CD"       = ~ n_perc(diagnosis_last_record=="CD", show_denom = "never"),
              "UC"       = ~ n_perc(diagnosis_last_record=="UC", show_denom = "never"),
              "IBDU" = ~ n_perc(diagnosis_last_record=="IBDU", show_denom = "never"))
  )

whole <- summary_table(metadata, table_summary)
whole
metadata$disease <- factor(metadata$disease, order = TRUE, 
                           levels = c("IBD", "control"))
by_disease_state <- summary_table(dplyr::group_by(metadata, disease), table_summary)  


```



```{r, echo = FALSE, results = "asis"}
print(by_disease_state,
      rtitle = "Summary Statistics",
      cnames = c("IBD patients (N = 355)", "Controls (N = 471)"))
```

```{r check-wd}
getwd()
```
